requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./compile_requirements.sh
#
dj-database-url==0.5.0
django-appconf==1.0.3     # via django-select2
django-filter==2.2.0
django-select2==7.0.3
django==2.2.1
djangorestframework==3.11.0
psycopg2==2.8.2
pytz==2019.1              # via django
six==1.12.0               # via django-appconf
sqlparse==0.3.0           # via django
